@@ -9,47 +9,60 @@
     "license": "MIT",
     "require": {
         "php": "^8.1.0",
+        "ext-xdebug": "*",
         "ext-gd": "*",
         "ext-json": "*",
         "ext-pdo": "*",
         "akuechler/laravel-geoly": "^1.0",
-        "doctrine/dbal": "^2.10",
-        "fideloper/proxy": "^4.0",
-        "fruitcake/laravel-cors": "^2.0",
+        "doctrine/dbal": "^3.2",
         "intervention/image": "^2.7",
-        "laravel/framework": "^8.0",
+        "laravel/framework": "^10.0",
         "laravel/helpers": "^1.4",
         "laravel/horizon": "^5.0",
         "laravel/legacy-factories": "^1.1",
+        "laravel/telescope": "^5.2",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^3.0",
-        "league/flysystem-aws-s3-v3": "^1.0",
-        "nexmo/client": "^2.0",
-        "nexmo/laravel": "^2.1",
+        "laravel/ui": "^4.0",
+        "laravel/vonage-notification-channel": "^3.0",
+        "league/flysystem-aws-s3-v3": "^3.0",
         "predis/predis": "^1.1",
+        "psr/http-factory": "^1.0",
         "pusher/pusher-php-server": "^7.0",
         "rtheunissen/guzzle-log-middleware": "^1.0",
-        "sentry/sentry-laravel": "^3.1",
-        "spatie/laravel-fractal": "^5.5",
-        "tymon/jwt-auth": "1.0.*",
+        "sentry/sentry-laravel": "^4.10",
+        "spatie/laravel-fractal": "^6.0",
+        "symfony/mailer": "^6.1",
+        "tymon/jwt-auth": "^2.0",
+        "vonage/client": "^4.0",
+        "vonage/vonage-laravel": "^1.0.3",
         "zircote/swagger-php": "^3.0"
     },
     "require-dev": {
         "beyondcode/laravel-dump-server": "^1.0",
+        "brianium/paratest": "^7.4",
+        "fakerphp/faker": "^1.23",
         "filp/whoops": "^2.0",
-        "fzaninotto/faker": "^1.4",
+        "larastan/larastan": "^2.9",
+        "laravel/pint": "^1.17",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^5.0",
-        "phpunit/phpunit": "^9.0"
+        "nunomaduro/collision": "^7.0",
+        "pestphp/pest": "^2.36",
+        "phpunit/phpunit": "^10.5"
     },
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "php-http/discovery": true,
+            "pestphp/pest-plugin": true
+        }
     },
     "extra": {
         "laravel": {
-            "dont-discover": []
+            "dont-discover": [
+                "laravel/telescope"
+            ]
         }
     },
     "autoload": {
@@ -67,7 +80,7 @@
             "app/Libraries/helpers.php"
         ]
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
@@ -79,6 +92,9 @@
         ],
         "post-create-project-cmd": [
             "@php artisan key:generate --ansi"
+        ],
+        "post-update-cmd": [
+            "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
         ]
     }
 }
