@@ -2,23 +2,30 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Responses\Response;
-use Illuminate\Foundation\Bus\DispatchesJobs;
-use Illuminate\Routing\Controller as BaseController;
-use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Response;
+use Illuminate\Routing\Controller as BaseController;
 
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
 
-    /**
-     * @var Response
-     */
-    protected $response;
+    protected Response $response;
 
     public function __construct()
     {
-        $this->response = new Response();
+        $this->response = new Response;
+    }
+
+    public function setJwtToken(string $token): JsonResponse
+    {
+        return response()->json([
+            'access_token' => $token,
+            'token_type' => 'bearer',
+            'expires_in' => now()->addYear()->timestamp,
+        ]);
     }
 }
