@@ -34,7 +34,7 @@ class CronDeleteReportedPets extends Command
     /**
      * Execute the console command.
      */
-    public function handle()
+    public function handle(): void
     {
         DeleteReportedPetsAfterXDaysJob::dispatch();
     }
