@@ -7,13 +7,7 @@ use Throwable;
 
 class CustomValidationException extends Exception
 {
-    /**
-     * CustomValidationException constructor.
-     *
-     * @param string $message
-     * @param Throwable|null $previous
-     */
-    public function __construct($message = 'Something was wrong', Throwable $previous = null)
+    public function __construct(string $message = 'Something was wrong', ?Throwable $previous = null)
     {
         parent::__construct($message, 422, $previous);
     }
