@@ -8,14 +8,7 @@ use Throwable;
 
 class ForbiddenException extends Exception
 {
-    /**
-     * CustomValidationException constructor.
-     *
-     * CustomException constructor.
-     * @param string $message
-     * @param Throwable|null $previous
-     */
-    public function __construct($message = 'Action not allowed!', Throwable $previous = null)
+    public function __construct(string $message = 'Action not allowed!', ?Throwable $previous = null)
     {
         parent::__construct($message, Response::HTTP_FORBIDDEN, $previous);
     }
