@@ -7,13 +7,7 @@ use Throwable;
 
 class UnauthorizedException extends AuthenticationException
 {
-    /**
-     * UnauthorizedException constructor.
-     *
-     * @param string $message
-     * @param Throwable|null $previous
-     */
-    public function __construct($message = 'Unauthorized', Throwable $previous = null)
+    public function __construct(string $message = 'Unauthorized', ?Throwable $previous = null)
     {
         parent::__construct($message, [], $previous);
     }
