@@ -3,57 +3,54 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Report;
 use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class AdminUserController extends Controller
 {
-    /**
-     * @param Request $request
-     * @return JsonResponse
-     */
-    public function index(Request $request)
+    public function index(Request $request): JsonResponse
     {
         $itemsPerPage = $request->length ?? 15;
         $page = $request->start ?? 1;
-        $users = User::orderByDesc('created_at')->withTrashed()->paginate($itemsPerPage, ['*'], 'page', $page);
-        return response()->json($users, 200);
+        $users = User::query()
+            ->withTrashed()
+            ->orderByDesc('created_at')
+            ->paginate($itemsPerPage, ['*'], 'page', $page);
+
+        return response()->json($users);
     }
 
-    /**
-     * @param Request $request
-     * @param User $user
-     * @return JsonResponse
-     */
-    public function show(Request $request, User $user)
+    public function show(Request $request, User $user): JsonResponse
     {
-        return response()->json($user, 200);
+        return response()->json($user);
     }
 
-    /**
-     * @param Request $request
-     * @param User $user
-     * @return JsonResponse
-     */
-    public function reportedPets(Request $request, User $user)
+    public function reportedPets(Request $request, User $user): JsonResponse
     {
         $itemsPerPage = $request->length ?? 15;
         $page = $request->start ?? 1;
-        $reportedPets = $user->reportedPets()->orderByDesc('created_at')->with('user')->withTrashed()->paginate($itemsPerPage, ['*'], 'page', $page);
-        return response()->json($reportedPets, 200);
+        $reportedPets = Report::withTrashed()
+            ->where('user_id', $user->id)
+            ->where('status', Report::STATUS_REPORTED)
+            ->orderByDesc('created_at')
+            ->with('user')
+            ->paginate($itemsPerPage, ['*'], 'page', $page);
+
+        return response()->json($reportedPets);
     }
 
-    /**
-     * @param Request $request
-     * @param User $user
-     * @return JsonResponse
-     */
-    public function userPets(Request $request, User $user)
+    public function userPets(Request $request, User $user): JsonResponse
     {
         $itemsPerPage = $request->length ?? 15;
         $page = $request->start ?? 1;
-        $reportedPets = $user->pets()->orderByDesc('created_at')->with('user')->withTrashed()->paginate($itemsPerPage, ['*'], 'page', $page);
-        return response()->json($reportedPets, 200);
+        $reportedPets = $user->pets()
+            ->withTrashed()
+            ->orderByDesc('created_at')
+            ->with('user')
+            ->paginate($itemsPerPage, ['*'], 'page', $page);
+
+        return response()->json($reportedPets);
     }
 }
