@@ -2,67 +2,51 @@
 
 namespace App\Http\Controllers;
 
-use App\Events\ChatNewMessageEvent;
-use App\Exceptions\UserHaveNoMissingPetsException;
+use App\Actions\Chat\ActivateUserChatAction;
+use App\Actions\Chat\CreateMessageAction;
+use App\Actions\Chat\DeleteUserChatAction;
+use App\Actions\Chat\GetUserChatAction;
 use App\Http\Requests\ChatCreateRequest;
-use App\Libraries\ChatService;
 use App\Models\Chat;
-use App\Models\ReportedPet;
-use App\Models\TrackedReportedPet;
-use App\Models\User;
+use App\Tasks\Chat\GetUserChatsTask;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class ChatController extends Controller
 {
-    /**
-     * @param Request $request
-     * @return JsonResponse
-     */
-    public function index(Request $request)
+    public function getUserChats(Request $request): JsonResponse
     {
-        $chatService = new ChatService($request);
-        $chats = $chatService->getUserChat();
+        $chats = app(GetUserChatsTask::class)->run();
 
-        return response()->json($chats, 200);
+        return response()->json($chats);
     }
 
-    /**
-     * @param Request $request
-     * @param TrackedReportedPet $trackedReportedPet
-     * @return JsonResponse
-     */
-    public function getChat(Request $request, TrackedReportedPet $trackedReportedPet)
+    public function getUserChat(Request $request, Chat $chat): JsonResponse
     {
-        $chatService = new ChatService($request);
-        $chat = $chatService->getFormattedChat($trackedReportedPet);
-        return response()->json($chat, 200);
+        $chat = app(GetUserChatAction::class)->run($chat);
+
+        return response()->json($chat);
     }
 
-    /**
-     * @param ChatCreateRequest $request
-     * @param TrackedReportedPet $trackedReportedPet
-     * @return JsonResponse
-     * @throws UserHaveNoMissingPetsException
-     */
-    public function store(ChatCreateRequest $request, TrackedReportedPet $trackedReportedPet)
+    public function createChatMessage(ChatCreateRequest $request, Chat $chat): JsonResponse
     {
-        /** @var User $user */
-        $user = $request->user();
-        if($trackedReportedPet->pet->status != ReportedPet::STATUS_MISSING) {
-            throw new UserHaveNoMissingPetsException();
-        }
+        $chat = app(CreateMessageAction::class)->run($request, $chat);
 
-        $chat = [
-            'message' => $request->message,
-            'sender_user_id' => $user->id,
-            'receiver_user_id' => $trackedReportedPet->reportedPet->user_id,
-            'reported_pet_id' => $trackedReportedPet->reported_pet_id,
-            'missing_pet_id' => $trackedReportedPet->pet_id,
-        ];
+        return response()->json($chat);
+    }
 
-        Chat::create($chat);
-        event(new ChatNewMessageEvent($trackedReportedPet, $chat));
-        return response()->json($chat, 200);
+    public function activateUserChat(Request $request, Chat $chat): Response
+    {
+        app(ActivateUserChatAction::class)->run($chat);
+
+        return response()->noContent();
+    }
+
+    public function deleteUserChat(Request $request, Chat $chat): Response
+    {
+        app(DeleteUserChatAction::class)->run($chat);
+
+        return response()->noContent();
     }
 }
