@@ -3,22 +3,23 @@
 namespace App\Http\Controllers;
 
 use App\Models\Pet;
-use App\Models\ReportedPet;
+use App\Models\Report;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class PublicController extends Controller
 {
-    public function statistics(Request $request)
+    public function statistics(Request $request): JsonResponse
     {
-        $totalUsers = User::all()->count();
+        $totalUsers = User::query()->count();
         $pets = Pet::all();
         $totalUserPets = $pets->count();
-        $totalUserMissingPets = $pets->where('status', ReportedPet::STATUS_MISSING)->count();
-        $reportedPets = ReportedPet::all();
-        $totalSpottedPets = $reportedPets->where('status', ReportedPet::STATUS_REPORTED)->count();
-        $totalFoundedPets = $reportedPets->where('status', ReportedPet::STATUS_FOUND)->count();
-        $totalFoundedPetsOnApp = $reportedPets->where('status', ReportedPet::STATUS_FOUND)->where('found_in_app', 1)->count();
+        $totalUserMissingPets = $pets->where('status', Report::STATUS_MISSING)->count();
+        $reportedPets = Report::all();
+        $totalSpottedPets = $reportedPets->where('status', Report::STATUS_REPORTED)->count();
+        $totalFoundedPets = $reportedPets->where('status', Report::STATUS_FOUND)->count();
+        $totalFoundedPetsOnApp = $reportedPets->where('status', Report::STATUS_FOUND)->where('found_in_app', 1)->count();
 
         $statistics = [
             'total_users' => $totalUsers,
@@ -28,6 +29,7 @@ class PublicController extends Controller
             'total_founded_pets' => $totalFoundedPets,
             'total_founded_pets_on_app' => $totalFoundedPetsOnApp,
         ];
-        return response()->json($statistics, 200);
+
+        return response()->json($statistics);
     }
 }
