@@ -3,20 +3,15 @@
 namespace App\Http\Middleware;
 
 use Illuminate\Auth\Middleware\Authenticate as Middleware;
-use Illuminate\Http\Request;
 
 class Authenticate extends Middleware
 {
-    /**
-     * Get the path the user should be redirected to when they are not authenticated.
-     *
-     * @param  Request  $request
-     * @return string
-     */
-    protected function redirectTo($request)
+    protected function redirectTo($request): string
     {
         if (! $request->expectsJson()) {
             return route('login');
         }
+
+        return '';
     }
 }
