@@ -3,28 +3,24 @@
 namespace App\Http\Middleware\User;
 
 use App\Models\User;
-use Carbon\Carbon;
 use Closure;
 use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 
 class CheckUserBannedMiddleware
 {
     /**
-     * Handle an incoming request.
-     *
-     * @param $request
-     * @param Closure $next
-     * @return mixed
      * @throws AuthorizationException
      */
-    public function handle($request, Closure $next)
+    public function handle(Request $request, Closure $next): JsonResponse
     {
         /** @var User $user */
         $user = auth()->user();
-        if($user->banned) {
-            if(Carbon::now() < $user->banned_at) {
-                $time = Carbon::now()->diffInMinutes($user->banned_at);
-                throw new AuthorizationException('You have been banned for '. $time.' minutes due to policy violation');
+        if ($user->banned) {
+            if (now() < $user->banned_at) {
+                $time = now()->diffInMinutes($user->banned_at);
+                throw new AuthorizationException('You have been banned for '.$time.' minutes due to policy violation');
             }
             $user->update([
                 'banned' => false,
