@@ -7,24 +7,13 @@ use Illuminate\Foundation\Http\FormRequest;
 class ChatCreateRequest extends FormRequest
 {
     /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
+     * @return string[]
      */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
+    public function rules(): array
     {
         return [
-            'message' => 'required',
+            'data' => 'required|max:255',
+            'receiver_id' => 'required|exists:users,id,deleted_at,NULL',
         ];
     }
 }
