@@ -7,24 +7,12 @@ use Illuminate\Foundation\Http\FormRequest;
 class UserEmailRequest extends FormRequest
 {
     /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
+     * @return string[]
      */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
+    public function rules(): array
     {
         return [
-            'email'     => 'required|email|unique:users,email',
+            'email' => 'required|email|unique:users,email',
         ];
     }
 }
