@@ -7,25 +7,12 @@ use Illuminate\Foundation\Http\FormRequest;
 class VerificationRequest extends FormRequest
 {
     /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
+     * @return string[]
      */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
+    public function rules(): array
     {
         return [
-            'phone'         => 'required|min:10|regex:/^[0-9\+]{7,15}$/',
-            'code_country'  => 'sometime',
+            'phone' => 'required|min:10|regex:/^[0-9\+]{7,15}$/',
         ];
     }
 }
