@@ -2,92 +2,93 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
 
-class Chat extends Model
+/**
+ * @property int $active
+ */
+class Chat extends ModelAbstract
 {
-    use SoftDeletes;
     const T = 'chats';
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
     protected $fillable = [
-        'message',
-        'sender_user_id',
-        'receiver_user_id',
-        'reported_pet_id',
-        'missing_pet_id',
+        'owner_id',
+        'reporter_id',
+        'report_found_id',
+        'report_missing_id',
+        'active',
     ];
 
-    /**
-     * The attributes that should be cast to native types.
-     *
-     * @var array
-     */
-    protected $casts = [];
-
-    protected $dates = [
-        'created_at',
-        'updated_at',
-        'deleted_at',
+    protected $hidden = [
+        'owner_id',
+        'reporter_id',
+        'report_found_id',
+        'report_missing_id',
     ];
 
-    public function toArray()
-    {
-        return [
-            'id'                => $this->id,
-            'message'           => $this->message,
-            'sender_user_id'    => $this->sender_user_id,
-            'receiver_user_id'  => $this->receiver_user_id,
-            'missing_pet_id'    => $this->missing_pet_id,
-            'created_at'        => $this->created_at,
-        ];
-    }
+    protected $casts = [
+        'active' => 'boolean',
+        'created_at' => 'datetime',
+        'updated_at' => 'datetime',
+        'deleted_at' => 'datetime',
+    ];
 
-//    public function getCreatedAtAttribute($value){
-//        $date = Carbon::parse($value);
-//        return $date->format('Y-m-d H:i');
-//    }
-//
-//    public function getUpdatedAtAttribute($value){
-//        $date = Carbon::parse($value);
-//        return $date->format('Y-m-d H:i');
-//    }
-
-    /**
-     * @return BelongsTo
-     */
-    public function senderUser()
+    public static function boot(): void
     {
-        return $this->belongsTo(User::class);
+        self::creating(function (Chat $model) {
+            $model->active = $model->active ?? 0;
+        });
+
+        parent::boot();
     }
 
     /**
-     * @return BelongsTo
+     * @return BelongsTo<User, Chat>
      */
-    public function receiverUser()
+    public function owner(): BelongsTo
     {
-        return $this->belongsTo(User::class);
+        return $this->belongsTo(User::class, 'owner_id', 'id');
     }
 
     /**
-     * @return BelongsTo
+     * @return BelongsTo<User, Chat>
      */
-    public function reportedPet()
+    public function reporter(): BelongsTo
     {
-        return $this->belongsTo(ReportedPet::class);
+        return $this->belongsTo(User::class, 'reporter_id', 'id');
     }
 
     /**
-     * @return BelongsTo
+     * @return BelongsTo<Report, Chat>
      */
-    public function missingPet()
+    public function reportFound(): BelongsTo
     {
-        return $this->belongsTo(Pet::class);
+        return $this->belongsTo(Report::class, 'report_found_id', 'id');
+    }
+
+    /**
+     * @return BelongsTo<Report, Chat>
+     */
+    public function reportMissing(): BelongsTo
+    {
+        return $this->belongsTo(Report::class, 'report_missing_id', 'id');
+    }
+
+    /**
+     * @return HasMany<Message>
+     */
+    public function messages(): HasMany
+    {
+        return $this->hasMany(Message::class, 'chat_id')->orderByDesc('id');
+    }
+
+    /**
+     * @return HasOne<Message>
+     */
+    public function lastMessage(): HasOne
+    {
+        return $this->hasOne(Message::class, 'chat_id')->latest();
     }
 }
