@@ -2,28 +2,17 @@
 
 namespace Database\Factories;
 
-use App\Models\User;
-use Illuminate\Database\Eloquent\Factory;
-use Illuminate\Support\Str;
-use Faker\Generator as Faker;
-
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
-/** @var Factory $factory */
-$factory->define(User::class, function (Faker $faker) {
-    return [
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'email_verified_at' => now(),
-        'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
-        'remember_token' => Str::random(10),
-    ];
-});
+class UserFactory extends FactoryAbstract
+{
+    public function definition(): array
+    {
+        return [
+            'name' => $this->faker->name,
+            'phone' => getPhoneNumber(),
+            //            'email' => $this->faker->unique()->safeEmail,
+            'email' => null,
+            'avatar' => null,
+            'email_verified_at' => now(),
+        ];
+    }
+}
