@@ -1,32 +1,14 @@
 <?php
 
-/*
-|--------------------------------------------------------------------------
-| Broadcast Channels
-|--------------------------------------------------------------------------
-|
-| Here you may register all of the event broadcasting channels that your
-| application supports. The given channel authorization callbacks are
-| used to check if an authenticated user can listen to the channel.
-|
-*/
+use App\Models\Pet;
 
-//Event: markedAs
-Broadcast::channel('pet', function ($user, $id) {
+//Event: getMissingPets
+Broadcast::channel('city.{city}.missingPets', function ($user, $id) {
     return (int) $user->id === (int) $id;
 });
 
-//Event: missing_pet
-Broadcast::channel('pet.{pet_id}.missing', function ($user, $id) {
-    return (int) $user->id === (int) $id;
-});
-
-//Event: sendMessage
-Broadcast::channel('chat.reportedPet.{reported_pet_id}.missingPet.{missing_pet_id}', function ($user, $id) {
-    return (int) $user->id === (int) $id;
-});
-
-//Event: sendMessage
-Broadcast::channel('chat.reportedPet.{reported_pet_id}.missingPet.{missing_pet_id}.typing', function ($user, $id) {
-    return (int) $user->id === (int) $id;
+//Event: reportCreated
+//Event: reportDeleted
+Broadcast::channel('{category}.{city}.report', function ($user) {
+    return $user->pets()->where('status', Pet::STATUS_MISSING)->exists();
 });
